requirements.txt
Django==5.1.4
pillow==11.1.0
gunicorn==22.0.0
django-cors-headers==4.6.0
qrcode==8.0
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
rest-framework-simplejwt==0.0.2
deepface==0.0.93
tf_keras==2.18.0
psycopg2-binary==2.9.10
python-decouple==3.8
openpyxl==3.1.5
